Install unzip to extract Composer archives
Summary:
Composer uses ZIP format for the archives. It tries to use native unzip utility,
with the ZipArchive PHP extension as fallback.

The preferred method is OS-dependant, with ZipArchive on Windows and unzip
everywhere else. The unzip utility is recommended for permissions or symlinks.

Reference: https://getcomposer.org/doc/articles/troubleshooting.md#zip-archives-are-not-unpacked-correctly

Ref T1455

Test Plan: Build image

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1455

Differential Revision: https://devcentral.nasqueron.org/D1813