Move services to /var/service
Summary:
In Debian Bullseye, the Debian package creates /etc/service as a symlink to
/etc/runit/runsvdir/current. The /etc/runit folder is a byzantine complex
construction with for example a "current" and "default" set of services.

This change switches to an independent /var/service folder, where images
can maintain a simpler structure.

This fixes issue to build the image using BuildKit, as this mechanism
doesn't allow to replace a symlink by a directory.

Ref T1679.

Test Plan: Build image with environment variable DOCKER_BUILDKIT set to 1.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1679

Differential Revision: https://devcentral.nasqueron.org/D2525